nginx-gen/requirements.txt
crossplane~=0.5.7
PyYAML~=6.0